refactor: 💡 move model hook to groups/index
✅ Closes: https://hashicorp.atlassian.net/browse/ICU-12591
